+++ packages/shaders/src/shaders/voronoi.ts
@@ -27,7 +27,7 @@ export const voronoiMeta = {
 
 // language=GLSL
 export const voronoiFragmentShader: string = `#version 300 es
-precision highp float;
+precision lowp float;
 
 uniform float u_time;
